Fixing vue

+++ app/Services/Server/ServerService.php
@@ -105,16 +105,21 @@ class ServerService implements ServerServiceContract
     /**
      * @param Server $server
      *
+     * @param bool $noDelete
      * @return mixed
      */
-    public function getStatus(Server $server)
+    public function getStatus(Server $server, $noDelete = false)
     {
         $server->touch();
 
         try {
-            return $this->getProvider($server->serverProvider)->getStatus($server);
+            $status = $this->getProvider($server->serverProvider)->getStatus($server);
+            $server->status = $status;
+            $server->save();
+
+            return $status;
         } catch (\Exception $e) {
-            if ($e->getMessage() == 'The resource you were accessing could not be found.') {
+            if (! $noDelete && $e->getMessage() == 'The resource you were accessing could not be found.') {
                 $server->delete();
 
                 return 'Server Has Been Deleted';
@@ -278,13 +283,12 @@ class ServerService implements ServerServiceContract
     /**
      * @param Server $server
      * @param $sshKey
-     * @param string $user
      *
      * @return bool
      */
-    public function installSshKey(Server $server, $sshKey, $user = 'root')
+    public function installSshKey(Server $server, $sshKey)
     {
-        $this->remoteTaskService->ssh($server, $user);
+        $this->remoteTaskService->ssh($server, 'codepier');
 
         $this->remoteTaskService->appendTextToFile('/home/codepier/.ssh/authorized_keys', $sshKey);
 
@@ -299,7 +303,7 @@ class ServerService implements ServerServiceContract
      */
     public function removeSshKey(Server $server, $sshKey)
     {
-        $this->remoteTaskService->ssh($server);
+        $this->remoteTaskService->ssh($server, 'codepier');
 
         $this->remoteTaskService->removeLineByText('/home/codepier/.ssh/authorized_keys', $sshKey);
